requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
plotly
pycountry
requests
certifi
python-dotenv